DR2-1739 Database builder and webapp
I've moved everything to a multiproject build and used sbt to build and
tag the docker images.
