+++ .scalafmt.conf
@@ -1,4 +1,4 @@
 version = 3.8.2
 preset = default
 runner.dialect = scala3
-maxColumn = 120
+maxColumn = 160
