+++ disaster-recovery-builder/src/main/resources/application.conf
@@ -1,0 +1,4 @@
+queue-url=${QUEUE_URL}
+database-path=${DATABASE_PATH}
+ocfl-repo-dir=${OCFL_REPO_DIR}
+ocfl-work-dir=${OCFL_WORK_DIR}
