+++ disaster-recovery-builder/src/main/resources/log4j2.properties
@@ -1,0 +1,11 @@
+status = warn
+
+name = DR2LambdaLogConfiguration
+
+appender.console.type = Console
+appender.console.name = consoleLogger
+appender.console.json.type = JsonTemplateLayout
+
+rootLogger.level = info
+
+rootLogger.appenderRef.stdout.ref = consoleLogger
