+++ disaster-recovery-builder/src/main/scala/uk/gov/nationalarchives/builder/Database.scala
@@ -1,0 +1,35 @@
+package uk.gov.nationalarchives.builder
+
+import cats.effect.Async
+import cats.implicits.*
+import uk.gov.nationalarchives.utils.Utils.OcflFile
+import doobie.Update
+import doobie.implicits.*
+import doobie.util.log.LogHandler
+import doobie.util.transactor.Transactor
+import doobie.util.transactor.Transactor.Aux
+import org.typelevel.log4cats.slf4j.Slf4jLogger
+
+trait Database[F[_]]:
+  def write(files: List[OcflFile]): F[Unit]
+
+object Database:
+  def apply[F[_]](using ev: Database[F]): Database[F] = ev
+
+  given impl[F[_]: Async](using configuration: Configuration): Database[F] = new Database[F] {
+
+    val xa: Aux[F, Unit] = Transactor.fromDriverManager[F](
+      driver = "org.sqlite.JDBC",
+      url = s"jdbc:sqlite:${configuration.config.databasePath}",
+      logHandler = Option(LogHandler.jdkLogHandler)
+    )
+
+    override def write(files: List[OcflFile]): F[Unit] = {
+      val sql = "insert into files (version, id, name, fileId, zref, path, fileName) values (?, ?, ?, ?, ?, ?, ?)"
+      for {
+        logger <- Slf4jLogger.create[F]
+        ret <- Update[OcflFile](sql).updateMany(files).transact(xa)
+        _ <- logger.info(s"Created $ret rows")
+      } yield ()
+    }
+  }
