+++ disaster-recovery-builder/src/test/resources/application.conf
@@ -1,0 +1,4 @@
+queue-url="http://localhost:9001"
+database-path="test-database"
+ocfl-repo-dir="/tmp/repo"
+ocfl-work-dir="/tmp/work"
