+++ disaster-recovery-webapp/src/main/resources/application.conf
@@ -1,0 +1,1 @@
+database-path=${DATABASE_PATH}
