+++ disaster-recovery-webapp/src/main/scala/uk/gov/nationalarchives/webapp/Main.scala
@@ -1,0 +1,7 @@
+package uk.gov.nationalarchives.webapp
+
+import cats.effect.{ExitCode, IO, IOApp}
+
+object Main extends IOApp:
+
+  override def run(args: List[String]): IO[ExitCode] = FrontEndServer.run[IO]
