+++ disaster-recovery-webapp/src/test/resources/application.conf
@@ -1,0 +1,1 @@
+database-path=test-database
