+++ disaster-recovery/src/main/scala/uk/gov/nationalarchives/disasterrecovery/Message.scala
@@ -1,0 +1,12 @@
+package uk.gov.nationalarchives.disasterrecovery
+
+import java.util.UUID
+
+sealed trait Message {
+  val ref: UUID
+  val messageText: String
+}
+object Message {
+  case class InformationObjectMessage(ref: UUID, messageText: String) extends Message
+  case class ContentObjectMessage(ref: UUID, messageText: String) extends Message
+}
