+++ project/plugins.sbt
@@ -2,4 +2,5 @@ addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "2.2.0")
 addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.5.2")
 addSbtPlugin("uk.gov.nationalarchives" % "sbt-assembly-log4j2" % "0.0.4")
 addSbtPlugin("com.github.sbt" % "sbt-release" % "1.4.0")
-addSbtPlugin("com.eed3si9n" % "sbt-buildinfo" % "0.12.0")
+addSbtPlugin("org.playframework.twirl" % "sbt-twirl" % "2.0.5")
+addSbtPlugin("com.github.sbt" % "sbt-native-packager" % "1.10.0")
