+++ utils/src/main/scala/uk/gov/nationalarchives/utils/Utils.scala
@@ -1,0 +1,4 @@
+package uk.gov.nationalarchives.utils
+
+object Utils:
+  final case class OcflFile(versionNum: Long, id: String, name: String, fileId: String, zref: String, path: String, fileName: String)
